feat: move init functions to game_init
