--- src/consumables.h
@@ -1,8 +1,0 @@
-#ifndef CONSUMABLES_H_
-#define CONSUMABLES_H_
-
-#include "game_core.h"
-
-void initConsumableBag(Fighter * f, Koliseo * kls);
-
-#endif // CONSUMABLES_H_
