+++ src/helapordo.c
@@ -762,486 +762,6 @@ void register_counter_callback(int index, callback_void_t ptr, Fighter* f) {
 }
 */
 
-//Status effect functions
-
-/**
- * Takes a Fighter pointer and resets its status value to Normal.
- * @see Fighter
- * @see fighterStatus
- * @param f The pointer whose status value will be reset.
- */
-void resetFighterStatus(Fighter *f)
-{
-    f->status = Normal;
-}
-
-/**
- * Takes a Enemy pointer and resets its status value to Normal.
- * @see Enemy
- * @see fighterStatus
- * @param e The pointer whose status value will be reset.
- */
-void resetEnemyStatus(Enemy *e)
-{
-    e->status = Normal;
-}
-
-/**
- * Takes a Boss pointer and resets its status value to Normal.
- * @see Boss
- * @see fighterStatus
- * @param b The pointer whose status value will be reset.
- */
-void resetBossStatus(Boss *b)
-{
-    b->status = Normal;
-}
-
-/**
- * Takes a fighterStatus and returns the corresponding effect_fun pointer for the function relating to the status.
- * @see fighterStatus
- * @see effect_fun
- * @param status The fighterStatus that the caller needs to match to a effect_fun pointer.
- * @return The effect_fun pointer related to the status.
- */
-effect_fun getStatusCounterFun(fighterStatus status)
-{
-    switch (status) {
-    //case Normal: {
-    default: {
-        return &resetFighterStatus;
-    }
-    break;
-    };
-}
-
-/**
- * Takes a fighterStatus and returns the corresponding effect_e_fun pointer for the function relating to the status.
- * @see fighterStatus
- * @see effect_e_fun
- * @param status The fighterStatus that the caller needs to match to a effect_e_fun pointer.
- * @return The effect_e_fun pointer related to the status.
- */
-effect_e_fun getStatusCounterEnemyFun(fighterStatus status)
-{
-    switch (status) {
-    //case Normal: {
-    default: {
-        return &resetEnemyStatus;
-    }
-    break;
-    };
-}
-
-/**
- * Takes a fighterStatus and returns the corresponding effect_e_fun pointer for the function relating to the status.
- * @see fighterStatus
- * @see effect_e_fun
- * @param status The fighterStatus that the caller needs to match to a effect_e_fun pointer.
- * @return The effect_e_fun pointer related to the status.
- */
-effect_b_fun getStatusCounterBossFun(fighterStatus status)
-{
-    switch (status) {
-    //case Normal: {
-    default: {
-        return &resetBossStatus;
-    }
-    break;
-    };
-}
-
-//Stat boost functions
-
-/**
- * Takes a Fighter pointer and an integer, used to set the turnboost_atk value at the pointer.
- * @see Fighter
- * @see Turncounter
- * @param f The Fighter pointer whose turnboost_atk value will be set.
- * @param boost The value that will be set.
- */
-void set_turnboost_atk(Fighter *f, int boost)
-{
-    //Straight addition of the boost
-    f->turnboost_atk = boost;
-}
-
-/**
- * Takes a Fighter pointer and an integer, used to set the turnboost_def value at the pointer.
- * @see Fighter
- * @see Turncounter
- * @param f The Fighter pointer whose turnboost_def value will be set.
- * @param boost The value that will be set.
- */
-void set_turnboost_def(Fighter *f, int boost)
-{
-    //Straight addition of the boost
-    f->turnboost_def = boost;
-}
-
-/**
- * Takes a Fighter pointer and an integer, used to set the turnboost_vel value at the pointer.
- * @see Fighter
- * @see Turncounter
- * @param f The Fighter pointer whose turnboost_vel value will be set.
- * @param boost The value that will be set.
- */
-void set_turnboost_vel(Fighter *f, int boost)
-{
-    //Straight addition of the boost
-    f->turnboost_vel = boost;
-}
-
-/**
- * Takes a Fighter pointer and an integer, used to set the turnboost_enr value at the pointer.
- * @see Fighter
- * @see Turncounter
- * @param f The Fighter pointer whose turnboost_enr value will be set.
- * @param boost The value that will be set.
- */
-void set_turnboost_enr(Fighter *f, int boost)
-{
-    //Straight addition of the boost
-    f->turnboost_enr = boost;
-}
-
-/**
- * Takes a Enemy pointer and an integer, used to set the turnboost_atk value at the pointer.
- * @see Enemy
- * @see Turncounter
- * @param e The Enemy pointer whose turnboost_atk value will be set.
- * @param boost The value that will be set.
- */
-void set_enemy_turnboost_atk(Enemy *e, int boost)
-{
-    //Straight addition of the boost
-    e->turnboost_atk = boost;
-}
-
-/**
- * Takes a Enemy pointer and an integer, used to set the turnboost_def value at the pointer.
- * @see Enemy
- * @see Turncounter
- * @param e The Enemy pointer whose turnboost_def value will be set.
- * @param boost The value that will be set.
- */
-void set_enemy_turnboost_def(Enemy *e, int boost)
-{
-    //Straight addition of the boost
-    e->turnboost_def = boost;
-}
-
-/**
- * Takes a Enemy pointer and an integer, used to set the turnboost_vel value at the pointer.
- * @see Enemy
- * @see Turncounter
- * @param e The Enemy pointer whose turnboost_vel value will be set.
- * @param boost The value that will be set.
- */
-void set_enemy_turnboost_vel(Enemy *e, int boost)
-{
-    //Straight addition of the boost
-    e->turnboost_vel = boost;
-}
-
-/**
- * Takes a Enemy pointer and an integer, used to set the turnboost_enr value at the pointer.
- * @see Enemy
- * @see Turncounter
- * @param e The Enemy pointer whose turnboost_enr value will be set.
- * @param boost The value that will be set.
- */
-void set_enemy_turnboost_enr(Enemy *e, int boost)
-{
-    //Straight addition of the boost
-    e->turnboost_enr = boost;
-}
-
-/**
- * Takes a Boss pointer and an integer, used to set the turnboost_atk value at the pointer.
- * @see Boss
- * @see Turncounter
- * @param b The Boss pointer whose turnboost_atk value will be set.
- * @param boost The value that will be set.
- */
-void set_boss_turnboost_atk(Boss *b, int boost)
-{
-    //Straight addition of the boost
-    b->turnboost_atk = boost;
-}
-
-/**
- * Takes a Boss pointer and an integer, used to set the turnboost_def value at the pointer.
- * @see Boss
- * @see Turncounter
- * @param b The Boss pointer whose turnboost_def value will be set.
- * @param boost The value that will be set.
- */
-void set_boss_turnboost_def(Boss *b, int boost)
-{
-    //Straight addition of the boost
-    b->turnboost_def = boost;
-}
-
-/**
- * Takes a Boss pointer and an integer, used to set the turnboost_vel value at the pointer.
- * @see Boss
- * @see Turncounter
- * @param b The Boss pointer whose turnboost_vel value will be set.
- * @param boost The value that will be set.
- */
-void set_boss_turnboost_vel(Boss *b, int boost)
-{
-    //Straight addition of the boost
-    b->turnboost_vel = boost;
-}
-
-/**
- * Takes a Boss pointer and an integer, used to set the turnboost_enr value at the pointer.
- * @see Boss
- * @see Turncounter
- * @param b The Boss pointer whose turnboost_enr value will be set.
- * @param boost The value that will be set.
- */
-void set_boss_turnboost_enr(Boss *b, int boost)
-{
-    //Straight addition of the boost
-    b->turnboost_enr = boost;
-}
-
-/**
- * Takes a FoeParty pointer and an integer, used to set the turnboost_atk value at the pointer.
- * @see FoeParty
- * @see Turncounter
- * @param fp The FoeParty pointer whose turnboost_atk value will be set.
- * @param boost The value that will be set.
- */
-void set_foeparty_turnboost_atk(FoeParty *fp, int boost)
-{
-    //Straight addition of the boost
-    fp->turnboost_atk = boost;
-}
-
-/**
- * Takes a FoeParty pointer and an integer, used to set the turnboost_def value at the pointer.
- * @see FoeParty
- * @see Turncounter
- * @param fp The FoeParty pointer whose turnboost_def value will be set.
- * @param boost The value that will be set.
- */
-void set_foeparty_turnboost_def(FoeParty *fp, int boost)
-{
-    //Straight addition of the boost
-    fp->turnboost_def = boost;
-}
-
-/**
- * Takes a FoeParty pointer and an integer, used to set the turnboost_vel value at the pointer.
- * @see FoeParty
- * @see Turncounter
- * @param fp The FoeParty pointer whose turnboost_vel value will be set.
- * @param boost The value that will be set.
- */
-void set_foeparty_turnboost_vel(FoeParty *fp, int boost)
-{
-    //Straight addition of the boost
-    fp->turnboost_vel = boost;
-}
-
-/**
- * Takes a FoeParty pointer and an integer, used to set the turnboost_enr value at the pointer.
- * @see FoeParty
- * @see Turncounter
- * @param fp The FoeParty pointer whose turnboost_enr value will be set.
- * @param boost The value that will be set.
- */
-void set_foeparty_turnboost_enr(FoeParty *fp, int boost)
-{
-    //Straight addition of the boost
-    fp->turnboost_atk = boost;
-}
-
-//Counter callback getters
-
-/**
- * Takes a Stat and returns the corresponding boost_fun pointer for the function relating to the stat.
- * @see Stat
- * @see boost_fun
- * @param s The Stat that the caller needs to match to a boost_fun pointer.
- * @return The boost_fun pointer related to the stat.
- */
-boost_fun getStatBoostCounterFun(Stat s)
-{
-
-    switch (s) {
-    case ATK: {
-        return &set_turnboost_atk;
-    }
-    break;
-    case DEF: {
-        return &set_turnboost_def;
-    }
-    break;
-    case VEL: {
-        return &set_turnboost_vel;
-    }
-    break;
-    case ENR: {
-        return &set_turnboost_enr;
-    }
-    break;
-    default: {
-        log_tag("debug_log.txt", "[ERROR]",
-                "Invalid stat value in getStatBoostCounterFun()");
-        exit(EXIT_FAILURE);
-    }
-    break;
-    };
-}
-
-/**
- * Takes a Stat and returns the corresponding boost_e_fun pointer for the function relating to the stat.
- * @see Stat
- * @see boost_e_fun
- * @param s The Stat that the caller needs to match to a boost_e_fun pointer.
- * @return The boost_e_fun pointer related to the stat.
- */
-boost_e_fun getStatBoostCounterEnemyFun(Stat s)
-{
-
-    switch (s) {
-    case ATK: {
-        return &set_enemy_turnboost_atk;
-    }
-    break;
-    case DEF: {
-        return &set_enemy_turnboost_def;
-    }
-    break;
-    case VEL: {
-        return &set_enemy_turnboost_vel;
-    }
-    break;
-    case ENR: {
-        return &set_enemy_turnboost_enr;
-    }
-    break;
-    default: {
-        log_tag("debug_log.txt", "[ERROR]",
-                "Invalid stat value in getStatBoostCounterEnemyFun()");
-        exit(EXIT_FAILURE);
-    }
-    break;
-    };
-}
-
-/**
- * Takes a Stat and returns the corresponding boost_b_fun pointer for the function relating to the stat.
- * @see Stat
- * @see boost_b_fun
- * @param s The Stat that the caller needs to match to a boost_b_fun pointer.
- * @return The boost_b_fun pointer related to the stat.
- */
-boost_b_fun getStatBoostCounterBossFun(Stat s)
-{
-
-    switch (s) {
-    case ATK: {
-        return &set_boss_turnboost_atk;
-    }
-    break;
-    case DEF: {
-        return &set_boss_turnboost_def;
-    }
-    break;
-    case VEL: {
-        return &set_boss_turnboost_vel;
-    }
-    break;
-    case ENR: {
-        return &set_boss_turnboost_enr;
-    }
-    break;
-    default: {
-        log_tag("debug_log.txt", "[ERROR]",
-                "Invalid stat value in getStatBoostCounterBossFun()");
-        exit(EXIT_FAILURE);
-    }
-    break;
-    };
-}
-
-/**
- * Takes a Stat and returns the corresponding boost_fp_fun pointer for the function relating to the stat.
- * @see Stat
- * @see boost_fp_fun
- * @param s The Stat that the caller needs to match to a boost_fp_fun pointer.
- * @return The boost_fp_fun pointer related to the stat.
- */
-boost_fp_fun get_StatBoostCounter_FoeParty_Fun(Stat s)
-{
-
-    switch (s) {
-    case ATK: {
-        return &set_foeparty_turnboost_atk;
-    }
-    break;
-    case DEF: {
-        return &set_foeparty_turnboost_def;
-    }
-    break;
-    case VEL: {
-        return &set_foeparty_turnboost_vel;
-    }
-    break;
-    case ENR: {
-        return &set_foeparty_turnboost_enr;
-    }
-    break;
-    default: {
-        log_tag("debug_log.txt", "[ERROR]",
-                "Invalid stat value in getStatBoostCounter_FoeParty_Fun()");
-        exit(EXIT_FAILURE);
-    }
-    break;
-    };
-}
-
-/**
- * Takes a Fighter pointer and prepares its perks field by allocating all the Perk accounted in perkClass.
- * Perks are initialised using the ordering of perkClass.
- * @see Fighter
- * @see Perk
- * @see perkClass
- * @see PERKSMAX
- * @param f The Fighter pointer whose perks field will be initialised.
- * @param kls The Koliseo used for allocation.
- */
-void initPerks(Fighter *f, Koliseo *kls)
-{
-    f->perksCount = 0;
-    //Ordering of i corresponds to perksClass enum
-    int total = (PERKSMAX + 1);
-    for (int i = 0; i < total; i++) {
-        kls_log(kls, "DEBUG", "Prepping Perk (%i)", i);
-        Perk *p =
-            (Perk *) KLS_PUSH_TYPED(kls, Perk, HR_Perk, "Perk", "Perk");
-        p->class = i;
-        char *name = nameStringFromPerk(i);
-        char *desc = descStringFromPerk(i);
-
-        //p->name = name; //(char*)malloc(sizeof(name));
-        strcpy(p->name, name);
-
-        //p->desc = (char*)malloc(sizeof(desc));
-        strcpy(p->desc, desc);
-
-        p->innerValue = 0;
-
-        f->perks[i] = p;
-    }
-}
-
 /**
  * Takes a Equip and a Fighter pointers.
  * Iterates over the equip's perks and removes them to the fighter perks.
@@ -1356,598 +876,6 @@ void printActivePerks(Fighter *f)
 
 //Counter functions
 
-/**
- * Takes a Fighter pointer and prepares its counters field by allocating all the Turncounter accounted in counterIndexes.
- * Counters are initialised using the ordering of counterIndexes, primarily for distinguinsing what kind of function pointer will be registered in the counter.
- * For a given Counter, only the correct pointer field is assigned and the others are left untouched.
- * @see Fighter
- * @see Turncounter
- * @see Countertype
- * @see COUNTERSMAX
- * @see counterIndexes
- * @see getStatBoostCounterFun()
- * @see getStatusCounterFun()
- * @param f The Fighter pointer whose counters field will be initialised.
- * @param kls The Koliseo used for allocation.
- */
-void initCounters(Fighter *f, Koliseo *kls)
-{
-    //Ordering of i corresponds to counterIndexes enum
-    int total = (COUNTERSMAX + 1);
-    for (int i = 0; i < total; i++) {
-        log_tag("debug_log.txt", "[DEBUG]", "Prepping Turncounter (%i)", i);
-        kls_log(kls, "DEBUG", "Prepping Turncounter (%i)", i);
-        Turncounter *c =
-            (Turncounter *) KLS_PUSH_TYPED(kls, Turncounter, HR_Turncounter,
-                                           "Turncounter", "Turncounter");
-
-        //First, prepare counters for statuses
-        if (i < STATUSMAX + 1) {
-            c->desc =
-                (char *)KLS_PUSH_ARR_TYPED(kls, char *, sizeof(stringFromStatus(i)),
-                                           HR_Turncounter_desc, "Turncounter desc",
-                                           "Turncounter desc");
-            strcpy(c->desc, stringFromStatus(i));
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "Allocated size %lu for status counter: (%s)",
-                    sizeof(stringFromStatus(i)), c->desc);
-            kls_log(kls, "DEBUG", "Allocated size %lu for status counter: (%s)",
-                    sizeof(stringFromStatus(i)), c->desc);
-
-            c->effect_fun = getStatusCounterFun(i);
-            //sprintf(msg,"Status function pointer is: (%i)", (int) *(c->effect_fun));
-            //log_tag("debug_log.txt","[DEBUG]",msg);
-            c->type = CNT_STATUS;
-
-            //Register function for the specific status counter
-            //
-            //REGISTER_COUNTER_CALLBACK(i,resetFighterStatus);
-            //
-            //Why does uncommenting this cause problems to special moves callback?
-            //More specific, ONLY MAGE MOVES
-            //debug
-            //printf("%s",stringFromStatus(j));
-        } else {		//Then, counters for boosts to (all?) stats
-
-            switch (i) {
-            case TURNBOOST_ATK: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_TYPED(kls, char *, sizeof("ATK boost"),
-                                               HR_Turncounter_desc,
-                                               "Turncounter desc",
-                                               "Turncounter desc");
-                strcpy(c->desc, "ATK boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("ATK boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("ATK boost"), c->desc);
-
-                c->boost_fun = getStatBoostCounterFun(ATK);
-                c->type = CNT_ATKBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            case TURNBOOST_DEF: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_TYPED(kls, char *, sizeof("DEF boost"),
-                                               HR_Turncounter_desc,
-                                               "Turncounter desc",
-                                               "Turncounter desc");
-                strcpy(c->desc, "DEF boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("DEF boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("DEF boost"), c->desc);
-
-                c->boost_fun = getStatBoostCounterFun(DEF);
-                c->type = CNT_DEFBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            case TURNBOOST_VEL: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_TYPED(kls, char *, sizeof("VEL boost"),
-                                               HR_Turncounter_desc,
-                                               "Turncounter desc",
-                                               "Turncounter desc");
-                strcpy(c->desc, "VEL boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("VEL boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("VEL boost"), c->desc);
-
-                c->boost_fun = getStatBoostCounterFun(VEL);
-                c->type = CNT_VELBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            case TURNBOOST_ENR: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_TYPED(kls, char *, sizeof("ENR boost"),
-                                               HR_Turncounter_desc,
-                                               "Turncounter desc",
-                                               "Turncounter desc");
-                strcpy(c->desc, "ENR boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("ENR boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("ENR boost"), c->desc);
-
-                c->boost_fun = getStatBoostCounterFun(ENR);
-                c->type = CNT_ENRBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            default: {
-                printf
-                ("\n\nERROR in initCounters(): unexpected i: %i value \n\n",
-                 i);
-                exit(EXIT_FAILURE);
-            }
-            }
-        }			//End else
-        c->count = 0;
-        c->innerValue = 0;
-        f->counters[i] = (struct Turncounter *)c;
-    };				//End for all counters
-}
-
-/**
- * Takes a Enemy pointer and prepares its counters field by allocating all the Turncounter accounted in counterIndexes.
- * Counters are initialised using the ordering of counterIndexes, primarily for distinguinsing what kind of function pointer will be registered in the counter.
- * For a given Counter, only the correct pointer field is assigned and the others are left untouched.
- * @see Enemy
- * @see Turncounter
- * @see Countertype
- * @see COUNTERSMAX
- * @see counterIndexes
- * @see getStatBoostCounterEnemyFun()
- * @see getStatusCounterEnemyFun()
- * @param e The Enemy pointer whose counters field will be initialised.
- * @param t_kls The Koliseo_Temp used for allocations.
- */
-void initECounters(Enemy *e, Koliseo_Temp *t_kls)
-{
-    //Ordering of i corresponds to counterIndexes enum
-    int total = (COUNTERSMAX + 1);
-    Koliseo *kls = t_kls->kls;
-    for (int i = 0; i < total; i++) {
-        log_tag("debug_log.txt", "[DEBUG]", "Prepping enemy Turncounter (%i)",
-                i);
-        kls_log(kls, "DEBUG", "Prepping enemy Turncounter (%i)", i);
-        Turncounter *c =
-            (Turncounter *) KLS_PUSH_T_TYPED(t_kls, Turncounter,
-                                             HR_Turncounter, "Turncounter",
-                                             "Turncounter");
-
-        //First, prepare counters for statuses
-        if (i < STATUSMAX + 1) {
-            c->desc =
-                (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                             sizeof(stringFromStatus(i)),
-                                             HR_Turncounter_desc,
-                                             "Turncounter desc",
-                                             "Turncounter desc");
-            strcpy(c->desc, stringFromStatus(i));
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "Allocated size %lu for enemy status counter: (%s)",
-                    sizeof(stringFromStatus(i)), c->desc);
-            kls_log(kls, "DEBUG",
-                    "Allocated size %lu for enemy status counter: (%s)",
-                    sizeof(stringFromStatus(i)), c->desc);
-
-            c->effect_e_fun = getStatusCounterEnemyFun(i);
-            //sprintf(msg,"[DEBUG]    Enemy status function pointer is: (%i)", *(c->effect_e_fun));
-            //log_tag("debug_log.txt","[DEBUG]",msg);
-            c->type = CNT_STATUS;
-
-        } else {		//Then, counters for boosts to (all?) stats
-
-            switch (i) {
-            case TURNBOOST_ATK: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                                 sizeof("ATK boost"),
-                                                 HR_Turncounter_desc,
-                                                 "Turncounter desc",
-                                                 "Turncounter desc");
-                strcpy(c->desc, "ATK boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("ATK boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("ATK boost"), c->desc);
-
-                c->boost_e_fun = getStatBoostCounterEnemyFun(ATK);
-                c->type = CNT_ATKBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            case TURNBOOST_DEF: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                                 sizeof("DEF boost"),
-                                                 HR_Turncounter_desc,
-                                                 "Turncounter desc",
-                                                 "Turncounter desc");
-                strcpy(c->desc, "DEF boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("DEF boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("DEF boost"), c->desc);
-
-                c->boost_e_fun = getStatBoostCounterEnemyFun(DEF);
-                c->type = CNT_DEFBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            case TURNBOOST_VEL: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                                 sizeof("VEL boost"),
-                                                 HR_Turncounter_desc,
-                                                 "Turncounter desc",
-                                                 "Turncounter desc");
-                strcpy(c->desc, "VEL boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("VEL boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("VEL boost"), c->desc);
-
-                c->boost_e_fun = getStatBoostCounterEnemyFun(VEL);
-                c->type = CNT_VELBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            case TURNBOOST_ENR: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                                 sizeof("ENR boost"),
-                                                 HR_Turncounter_desc,
-                                                 "Turncounter desc",
-                                                 "Turncounter desc");
-                strcpy(c->desc, "ENR boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("ENR boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("ENR boost"), c->desc);
-
-                c->boost_e_fun = getStatBoostCounterEnemyFun(ENR);
-                c->type = CNT_ENRBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            default: {
-                log_tag("debug_log.txt", "[ERROR]",
-                        "ERROR INITALISING TURNCOUNTER in initECounters()\n");
-                exit(EXIT_FAILURE);
-            }
-            }
-        }			//End else
-        c->count = 0;
-        c->innerValue = 0;
-        e->counters[i] = (struct Turncounter *)c;
-    };
-}
-
-/**
- * Takes a Boss pointer and prepares its counters field by allocating all the Turncounter accounted in counterIndexes.
- * Counters are initialised using the ordering of counterIndexes, primarily for distinguinsing what kind of function pointer will be registered in the counter.
- * For a given Counter, only the correct pointer field is assigned and the others are left untouched.
- * @see Boss
- * @see Turncounter
- * @see Countertype
- * @see COUNTERSMAX
- * @see counterIndexes
- * @see getStatBoostCounterBossFun()
- * @see getStatusCounterBossFun()
- * @param b The Boss pointer whose counters field will be initialised.
- * @param t_kls The Koliseo_Temp used for allocations.
- */
-void initBCounters(Boss *b, Koliseo_Temp *t_kls)
-{
-    //Ordering of i corresponds to counterIndexes enum
-    int total = (COUNTERSMAX + 1);
-    Koliseo *kls = t_kls->kls;
-    for (int i = 0; i < total; i++) {
-        log_tag("debug_log.txt", "[DEBUG]", "Prepping boss counter %i", i);
-        kls_log(kls, "DEBUG", "Prepping boss counter %i", i);
-        Turncounter *c =
-            (Turncounter *) KLS_PUSH_T_TYPED(t_kls, Turncounter,
-                                             HR_Turncounter, "Turncounter",
-                                             "Turncounter");
-
-        //First, prepare counters for statuses
-        if (i < STATUSMAX + 1) {
-            c->desc =
-                (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                             sizeof(stringFromStatus(i)),
-                                             HR_Turncounter_desc,
-                                             "Turncounter desc",
-                                             "Turncounter desc");
-            strcpy(c->desc, stringFromStatus(i));
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "Allocated size %lu for boss status counter: (%s)",
-                    sizeof(stringFromStatus(i)), c->desc);
-            kls_log(kls, "DEBUG",
-                    "Allocated size %lu for boss status counter: (%s)",
-                    sizeof(stringFromStatus(i)), c->desc);
-
-            c->effect_b_fun = getStatusCounterBossFun(i);
-            //sprintf(msg,"Boss status function pointer is: (%i)", *(c->effect_b_fun));
-            //log_tag("debug_log.txt","[DEBUG]",msg);
-            c->type = CNT_STATUS;
-
-        } else {		//Then, counters for boosts to (all?) stats
-
-            switch (i) {
-            case TURNBOOST_ATK: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                                 sizeof("ATK boost"),
-                                                 HR_Turncounter_desc,
-                                                 "Turncounter desc",
-                                                 "Turncounter desc");
-                strcpy(c->desc, "ATK boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for boss stat counter: (%s)",
-                        sizeof("ATK boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for boss stat counter: (%s)",
-                        sizeof("ATK boost"), c->desc);
-
-                c->boost_b_fun = getStatBoostCounterBossFun(ATK);
-                c->type = CNT_ATKBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_b_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            case TURNBOOST_DEF: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                                 sizeof("DEF boost"),
-                                                 HR_Turncounter_desc,
-                                                 "Turncounter desc",
-                                                 "Turncounter desc");
-                strcpy(c->desc, "DEF boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for boss stat counter: (%s)",
-                        sizeof("DEF boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for boss stat counter: (%s)",
-                        sizeof("DEF boost"), c->desc);
-
-                c->boost_b_fun = getStatBoostCounterBossFun(DEF);
-                c->type = CNT_DEFBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_b_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            case TURNBOOST_VEL: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                                 sizeof("VEL boost"),
-                                                 HR_Turncounter_desc,
-                                                 "Turncounter desc",
-                                                 "Turncounter desc");
-                strcpy(c->desc, "VEL boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for boss stat counter: (%s)",
-                        sizeof("VEL boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for boss stat counter: (%s)",
-                        sizeof("VEL boost"), c->desc);
-
-                c->boost_b_fun = getStatBoostCounterBossFun(VEL);
-                c->type = CNT_VELBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_b_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            case TURNBOOST_ENR: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                                 sizeof("ENR boost"),
-                                                 HR_Turncounter_desc,
-                                                 "Turncounter desc",
-                                                 "Turncounter desc");
-                strcpy(c->desc, "ENR boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for boss stat counter: (%s)",
-                        sizeof("ENR boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for boss stat counter: (%s)",
-                        sizeof("ENR boost"), c->desc);
-
-                c->boost_b_fun = getStatBoostCounterBossFun(ENR);
-                c->type = CNT_ENRBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_b_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            default: {
-                log_tag("debug_log.txt", "[ERROR]",
-                        "ERROR INITALISING TURNCOUNTER in initBCounters()\n");
-                exit(EXIT_FAILURE);
-            }
-            }
-        }			//End else
-        c->count = 0;
-        c->innerValue = 0;
-        b->counters[i] = (struct Turncounter *)c;
-    };
-}
-
-/**
- * Takes a FoeParty pointer and prepares its counters field by allocating all the Turncounter accounted in counterIndexes.
- * Counters are initialised using the ordering of counterIndexes, primarily for distinguinsing what kind of function pointer will be registered in the counter.
- * For a given Counter, only the correct pointer field is assigned and the others are left untouched.
- * @see FoeParty
- * @see Turncounter
- * @see Countertype
- * @see COUNTERSMAX
- * @see counterIndexes
- * @see get_StatBoostCounter_FoeParty_Fun()
- * @see get_StatusCounter_FoeParty_Fun()
- * @param fp The FoeParty pointer whose counters field will be initialised.
- * @param t_kls The Koliseo_Temp used for allocations.
- */
-void initFoePartyCounters(FoeParty *fp, Koliseo_Temp *t_kls)
-{
-    //Ordering of i corresponds to counterIndexes enum
-    int total = (COUNTERSMAX + 1);
-    Koliseo *kls = t_kls->kls;
-    for (int i = 0; i < total; i++) {
-        log_tag("debug_log.txt", "[DEBUG]", "Prepping foeparty counter %i", i);
-        kls_log(kls, "DEBUG", "Prepping foeparty counter %i", i);
-        Turncounter *c =
-            (Turncounter *) KLS_PUSH_T_TYPED(t_kls, Turncounter,
-                                             HR_Turncounter, "Turncounter",
-                                             "Turncounter");
-
-        //First, prepare counters for statuses
-        if (i < STATUSMAX + 1) {
-            c->desc =
-                (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                             sizeof(stringFromStatus(i)),
-                                             HR_Turncounter_desc,
-                                             "Turncounter desc",
-                                             "Turncounter desc");
-            strcpy(c->desc, stringFromStatus(i));
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "Allocated size %lu for foeparty status counter: (%s)",
-                    sizeof(stringFromStatus(i)), c->desc);
-            kls_log(kls, "DEBUG",
-                    "Allocated size %lu for foeparty status counter: (%s)",
-                    sizeof(stringFromStatus(i)), c->desc);
-
-            //c->effect_fp_fun = get_StatusCounter_FoeParty_Fun(i);
-            //sprintf(msg,"Foeparty status function pointer is: (%i)", *(c->effect_b_fun));
-            //log_tag("debug_log.txt","[DEBUG]",msg);
-            c->type = CNT_STATUS;
-
-        } else {		//Then, counters for boosts to (all?) stats
-
-            switch (i) {
-            case TURNBOOST_ATK: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                                 sizeof("ATK boost"),
-                                                 HR_Turncounter_desc,
-                                                 "Turncounter desc",
-                                                 "Turncounter desc");
-                strcpy(c->desc, "ATK boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("ATK boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("ATK boost"), c->desc);
-
-                c->boost_fp_fun = get_StatBoostCounter_FoeParty_Fun(ATK);
-                c->type = CNT_ATKBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_fp_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            case TURNBOOST_DEF: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                                 sizeof("DEF boost"),
-                                                 HR_Turncounter_desc,
-                                                 "Turncounter desc",
-                                                 "Turncounter desc");
-                strcpy(c->desc, "DEF boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("DEF boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("DEF boost"), c->desc);
-
-                c->boost_fp_fun = get_StatBoostCounter_FoeParty_Fun(DEF);
-                c->type = CNT_DEFBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_fp_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            case TURNBOOST_VEL: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                                 sizeof("VEL boost"),
-                                                 HR_Turncounter_desc,
-                                                 "Turncounter desc",
-                                                 "Turncounter desc");
-                strcpy(c->desc, "VEL boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("VEL boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("VEL boost"), c->desc);
-
-                c->boost_fp_fun = get_StatBoostCounter_FoeParty_Fun(VEL);
-                c->type = CNT_VELBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_fp_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            case TURNBOOST_ENR: {
-                c->desc =
-                    (char *)KLS_PUSH_ARR_T_TYPED(t_kls, char *,
-                                                 sizeof("ENR boost"),
-                                                 HR_Turncounter_desc,
-                                                 "Turncounter desc",
-                                                 "Turncounter desc");
-                strcpy(c->desc, "ENR boost");
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("ENR boost"), c->desc);
-                kls_log(kls, "DEBUG",
-                        "Allocated size %lu for status counter: (%s)",
-                        sizeof("ENR boost"), c->desc);
-
-                c->boost_fp_fun = get_StatBoostCounter_FoeParty_Fun(ENR);
-                c->type = CNT_ENRBOOST;
-                //sprintf(msg,"Stat boost function pointer is: (%i)", *(c->boost_fp_fun));
-                //log_tag("debug_log.txt","[DEBUG]",msg);
-            }
-            break;
-            default: {
-                log_tag("debug_log.txt", "[ERROR]",
-                        "Invalid counter in initFoePartyCounters");
-                exit(EXIT_FAILURE);
-            }
-            }
-        }			//End else
-        c->count = 0;
-        c->innerValue = 0;
-        fp->counters[i] = (struct Turncounter *)c;
-    };
-}
 
 //TODO update this to actually handle different types of counters and reliably print them, maybe to a ncurses window
 /**
@@ -2125,39 +1053,6 @@ void setCounter(Turncounter *c, int turns)
     } else {
         //Handle counters already activ
     }
-}
-
-/**
- * Takes a Fighter pointer and prepares its skillSlot fields by allocating FIGHTER_SKILL_SLOTS slots.
- * Skill slots are initialised.
- * @see Fighter
- * @see Skilllot
- * @see FIGHTER_SKILL_SLOTS
- * @see SKILL_TYPE_LAST_UNLOCKABLE
- * @see costFromSkill()
- * @see stringFromSkill()
- * @param kls The Koliseo used for allocations.
- * @param f The Fighter pointer whose skill slots will be initialised.
- */
-void setSkills(Fighter *f, Koliseo *kls)
-{
-    char movename[80];
-    char movedesc[80];
-    for (int i = 0; i <= FIGHTER_SKILL_SLOTS; i++) {
-        kls_log(kls, "DEBUG", "Prepping Skillslot (%i)", i);
-        Skillslot *s =
-            (Skillslot *) KLS_PUSH_TYPED(kls, Skillslot, HR_Skillslot,
-                                         "Skillslot", "Skillslot");
-        s->enabled = 0;
-        s->class = i;
-        s->cost = costFromSkill(i);
-        strcpy(movename, nameStringFromSkill(i));
-        strcpy(movedesc, descStringFromSkill(i));
-        //printf("DEBUG\n%i\t%s\n",(i),stringFromSkill(i));
-        strcpy(s->name, movename);
-        strcpy(s->desc, movedesc);
-        f->skills[i] = s;
-    };
 }
 
 /**
@@ -2370,372 +1265,6 @@ void applyArtifacts(Fighter *f, Enemy *e, Boss *b, int isBoss)
             };
         };
     };
-}
-
-/**
- * Takes a Fighter pointer and prepares its equipslots field by allocating an Equipslot for each Equipzone.
- * @see Fighter
- * @see Equipslot
- * @see Equipzone
- * @see counterIndexes
- * @see EQUIPZONES
- * @param f The Fighter pointer whose equipslots field will be initialised.
- * @param kls The Koliseo used for allocations.
- */
-void initEquipSlots(Fighter *f, Koliseo *kls)
-{
-    for (int i = 0; i <= EQUIPZONES; i++) {
-        kls_log(kls, "DEBUG", "Prepping Equipslot (%i)", i);
-        Equipslot *s =
-            (Equipslot *) KLS_PUSH_TYPED(kls, Equipslot, HR_Equipslot,
-                                         "Equipslot", "Equipslot");
-        s->active = 0;
-        s->type = i;
-        setEquipslotSprite(s);
-        f->equipslots[i] = (struct Equipslot *)s;
-    };
-}
-
-/**
- * Takes a Fighter pointer and prepares its artifactsBag field by allocating a Artifact for each artifactClass.
- * @see Fighter
- * @see Artifact
- * @see artifactClass
- * @see ARTIFACTSMAX
- * @param kls The Koliseo used for allocations.
- * @param f The Fighter pointer whose artifactsBag field will be initialised.
- */
-void initArtifactsBag(Fighter *f, Koliseo *kls)
-{
-    for (int i = 0; i < ARTIFACTSMAX + 1; i++) {
-        kls_log(kls, "DEBUG", "Prepping Artifact (%i)", i);
-        Artifact *a =
-            (Artifact *) KLS_PUSH_TYPED(kls, Artifact, HR_Artifact,
-                                        "Artifact", "Artifact");
-        a->class = i;
-
-        Artifact *base = &artifactsBase[i];
-
-        strcpy(a->name, base->name);
-        strcpy(a->desc, base->desc);
-        for (int j = 0; j < 8; j++) {
-            strcpy(a->sprite[j], base->sprite[j]);
-        }
-        a->qty = 0;
-
-        f->artifactsBag[i] = (struct Artifact *)a;
-    }
-}
-
-/**
- * Takes one Fighter and one Path pointers and initialises the fighter fields.
- * Luck value is set as path luck value modulo MAXPLAYERLUCK.
- * The BaseStats pointer for the fighter's figtherClass is loaded.
- * The stats field is initalised with all inner values at 0.
- * setSpecials(), initCounters() are called to init the specialslots and counters fields.
- * initEquipSlots() is called to init the equipslots field, and the relative int field are set to 0.
- * initArtifactsFun() is called to init all the Artifact effect_fun field.
- * All the tempboost_STAT and permboost_STAT fields are set to 0.
- * All the fields common to BaseStats are initialised with the base value.
- * The status field is set to Normal.
- * @see Fighter
- * @see fighterClass
- * @see Path
- * @see BaseStats
- * @see countStats
- * @see MAXPLAYERLUCK
- * @see setSpecials()
- * @see initCounters()
- * @see initEquipSlots()
- * @see initArtifactsFun()
- * @see fighterStatus
- * @param player The Fighter whose fields will be initialised.
- * @param path The Path pointer of the current game.
- * @param kls The Koliseo used for allocation.
- */
-void initPlayerStats(Fighter *player, Path *path, Koliseo *kls)
-{
-
-    //player luck : MAXPLAYERLUCK = path luck : MAXLUCK
-
-    player->luck = (path->luck * MAXPLAYERLUCK) / MAXLUCK;
-
-    BaseStats *base = &basestats[player->class];
-
-    kls_log(kls, "DEBUG", "Prepping countStats");
-    countStats *s =
-        (countStats *) KLS_PUSH_TYPED(kls, countStats, HR_countStats,
-                                      "countStats", "countStats");
-
-    s->enemieskilled = 0;
-    s->criticalhits = 0;
-    s->consumablesfound = 0;
-    s->equipsfound = 0;
-    s->artifactsfound = 0;
-    s->roomscompleted = 0;
-    s->floorscompleted = 0;
-    s->specialsunlocked = 0;
-    s->coinsfound = 0;
-    s->bosseskilled = 0;
-    s->unique_bosseskilled = 0;
-    for (int i = 0; i < BOSSCLASSESMAX + 1; i++) {
-        s->killed_bosses[i] = 0;
-    }
-    s->keysfound = 0;
-
-    setSpecials(player, kls);
-    setSkills(player, kls);
-    initCounters(player, kls);
-    initPerks(player, kls);
-
-    initConsumableBag(player, kls);
-    initArtifactsBag(player, kls);
-
-    initEquipSlots(player, kls);
-    player->equipsBagOccupiedSlots = 0;	//Keeps track of how many slots are occupied.
-    player->earliestBagSlot = 0;	//To alwasy use the array efficiently (???) I sense linked lists may be better
-
-    player->permboost_atk = 0;
-    player->permboost_def = 0;
-    player->permboost_vel = 0;
-    player->permboost_enr = 0;
-
-    player->equipboost_atk = 0;
-    player->equipboost_def = 0;
-    player->equipboost_vel = 0;
-    player->equipboost_enr = 0;
-
-    player->balance = 0;
-    player->keys_balance = 0;
-
-    player->stats = s;
-    player->hp = base->hp;
-    player->totalhp = base->totalhp;
-    player->atk = base->atk;
-    player->def = base->def;
-    player->vel = base->vel;
-    player->level = base->level;
-    player->totalxp = base->totalxp;
-    player->currentlevelxp = base->currentlevelxp;
-    player->totallevelxp = base->totallevelxp;
-    player->status = Normal;
-    player->totalenergy = base->totalenergy;
-    player->energy = player->totalenergy;
-    player->totalstamina = base->totalstamina;
-    player->stamina = player->totalstamina;
-
-    setFighterSprite(player);
-}
-
-/**
- * Takes one Enemy pointer and initialises the enemy fields.
- * The EnemyBaseStats pointer for the enemy's enemyClass is loaded.
- * If the beast field at the pointer is already set before this call, atk def vel and hp of the enemy will receive a multiplicative boost equal to BSTFACTOR . Also, Xp value will be multiplied by 3.
- * Notably, the level field is not set at all by this function and needs to be set by the caller.
- * All the fields common to EnemyBaseStats are initialised with the base value and eventual beast boost.
- * initECounters() is called to init the counters field.
- * The status field is set to Normal.
- * @see Enemy
- * @see enemyClass
- * @see EnemyBaseStats
- * @see initECounters()
- * @see fighterStatus
- * @param e The Enemy whose fields will be initialised.
- * @param t_kls The Koliseo_Temp used for allocations.
- */
-void initEnemyStats(Enemy *e, Koliseo_Temp *t_kls)
-{
-    EnemyBaseStats *base = &baseenemystats[e->class];
-    log_tag("debug_log.txt", "[DEBUG]", "Init stats for enemy (%s)",
-            stringFromEClass(e->class));
-
-    float beastf = 1;
-
-    if (e->beast) {
-        beastf = BSTFACTOR;
-    }
-
-    e->hp = round(beastf * base->hp);
-    e->totalhp = round(beastf * base->totalhp);
-    e->atk = round(beastf * base->atk);
-    e->def = round(beastf * base->def);
-    e->vel = round(beastf * base->vel);
-    e->status = Normal;
-
-    //Level should be set by caller
-    //Index should be set by caller
-
-    e->totalenergy = base->totalenergy;
-    e->energy = e->totalenergy;
-    e->totalstamina = base->totalstamina;
-    e->stamina = e->totalstamina;
-
-    //Set prize, double for beasts
-    float prize = 2.8 * e->level;
-
-    e->prize = floor((e->beast) ? 2 * prize : prize);
-
-    initECounters(e, t_kls);
-
-    //Triple xp for beasts
-    e->xp = (e->beast) ? 3 * base->xp : base->xp;
-
-    setEnemySprite(e);
-}
-
-/**
- * Takes one Boss pointer and initialises the boss fields.
- * The BossBaseStats pointer for the boss's bossClass is loaded.
- * If the beast field at the pointer is already set before this call, atk def vel and hp of the enemy will receive a multiplicative boost equal to BSTFACTOR . Xp value will also be multiplie by 3.
- * Notably, the level field is not set at all by this function and needs to be set by the caller.
- * All the fields common to BossBaseStats are initialised with the base value and eventual beast boost.
- * initBCounters() is called to init the counters field.
- * The status field is set to Normal.
- * @see Boss
- * @see bossClass
- * @see BossBaseStats
- * @see initBCounters()
- * @see fighterStatus
- * @param b The Boss whose fields will be initialised.
- * @param t_kls The Koliseo_Temp used for allocations.
- */
-void initBossStats(Boss *b, Koliseo_Temp *t_kls)
-{
-    //Class should be set by caller
-    BossBaseStats *base = &basebossstats[b->class];
-
-    float beastf = 1;
-
-    if (b->beast) {
-        beastf = BSTFACTOR;
-    }
-
-    b->hp = round(beastf * base->hp);
-    b->totalhp = round(beastf * base->totalhp);
-    b->atk = round(beastf * base->atk);
-    b->def = round(beastf * base->def);
-    b->vel = round(beastf * base->vel);
-    b->status = Normal;
-
-    //Level should be set by caller
-
-    b->totalenergy = base->totalenergy;
-    b->energy = b->totalenergy;
-
-    b->totalstamina = base->totalstamina;
-    b->stamina = b->totalstamina;
-
-    //Set prize, double for beasts
-    float prize = 4.5 * b->level;
-
-    b->prize = (b->beast) ? 2 * prize : prize;
-
-    initBCounters(b, t_kls);
-
-    //Triple xp for beasts
-    b->xp = (b->beast) ? 3 * base->xp : base->xp;
-
-    setBossSprite(b);
-}
-
-/**
- * Takes one FoeParty pointer and initialises the party fields.
- * The FoePartyBaseStats pointer for the FoeParty's foePartyClass is loaded.
- * All the fields common to FoePartyBaseStats are initialised with the base value.
- * initFoePartyCounters() is called to init the counters field.
- * @see FoeParty
- * @see foePartyClass
- * @see FoePartyBaseStats
- * @see initFoePartyCounters()
- * @param fp The FoeParty whose fields will be initialised.
- * @param t_kls The Koliseo_Temp used for allocations.
- */
-void initFoePartyStats(FoeParty *fp, Koliseo_Temp *t_kls)
-{
-    //Class should be set by caller
-    //FoePartyBaseStats* base = &basefoepartystats[fp->class];
-    //TODO: add loading of more base stats not expected in prepareFoeParty()
-
-    //Level should be set by caller
-
-    //Zero the alive_enemies values
-    for (int i = 0; i < ROOM_ENEMIES_MAX + 1; i++) {
-        fp->alive_enemies[i] = 0;
-    };
-    //Zero the alive_bosses values
-    for (int i = 0; i < FOES_BOSSES_MAX + 1; i++) {
-        fp->alive_bosses[i] = 0;
-    };
-
-    //Zero the turnboost_STAT values
-    fp->turnboost_atk = 0;
-    fp->turnboost_def = 0;
-    fp->turnboost_vel = 0;
-    fp->turnboost_enr = 0;
-
-    //Zero current enemy index
-    fp->current_index = 0;
-
-    //Size value should be set by caller
-    fp->tot_alive = fp->size;
-
-    switch (fp->class) {
-    case Enemies: {
-        //Set alive_enemies values
-        for (int i = 0; i < fp->tot_alive + 1; i++) {
-            fp->alive_enemies[i] = 1;
-        };
-    }
-    break;
-    case Bosses: {
-        //Set alive_bosses values
-        for (int i = 0; i < fp->tot_alive + 1; i++) {
-            fp->alive_bosses[i] = 1;
-        };
-    }
-    break;
-    default: {
-        fprintf(stderr, "UNEXPECTED FOEPARTY CLASS VALUE %i", fp->class);
-        log_tag("debug_log.txt", "[PANIC]",
-                "Invalid class value in initFoePartyStats()");
-        exit(EXIT_FAILURE);
-    }
-    break;
-    };
-
-    //Init party counters
-    initFoePartyCounters(fp, t_kls);
-}
-
-/**
- * Takes a FoeParty pointer, a size for complete party, and an integer for the current room index, and initialises all the fields making the FoeParty ready for use in battle.
- * Calls initFoePartyStats() after setting class and level
- * @see FoeParty
- * @see initFoePartyStats()
- * @param fp The allocated FoeParty pointer to initialise.
- * @param t_kls The Koliseo_Temp used for allocations.
- */
-void prepareFoeParty(FoeParty *fp, int partysize, int roomindex,
-                     Koliseo_Temp *t_kls)
-{
-
-    //Class must be set by caller
-
-    FoePartyBaseStats *base = &basefoepartystats[fp->class];
-
-    fp->level = base->level;
-    //Set size value
-    fp->size = partysize;
-
-    //FoeParty's get 1 level every 2 rooms
-    //fp->level += floor(roomindex / 2) ;
-
-    //Load foeparty stats
-    initFoePartyStats(fp, t_kls);
-
-    log_tag("debug_log.txt", "[DEBUG]",
-            "Prepared FoeParty with size (%i), room #%i.", partysize,
-            roomindex);
 }
 
 /**
@@ -2995,567 +1524,6 @@ void prepareRoomEnemy(Enemy *e, int roomindex, int enemiesInRoom,
 
     //Force load of level bonuses
     statResetEnemy(e, 1);
-}
-
-/**
- * Takes a size int and a pointer to integer array, initialises all the prices, depending on stats from the Equips array pointer.
- * @see initShop()
- * @see Equip
- * @param size Size of both arrays.
- * @param equipPrices Pointer to integer array for prices.
- * @param equips Pointer to Equip pointer to influence pricing.
- */
-void setEquipPrices(int size, int *equipPrices, Equip *equips[])
-{
-
-    for (int i = 0; i < size; i++) {
-        int *cur_price = &equipPrices[i];
-        Equip *e = equips[i];
-
-        int baseprice = e->cost;
-        int price = baseprice;
-
-        *cur_price = price;
-    }
-
-}
-
-/**
- * Takes a size int and a pointer to integer array, initialises all the prices.
- * @see initShop()
- * @param size Size of both arrays.
- * @param consumablePrices Pointer to integer array for prices.
- * @param consumables Pointer to Consumable pointer to influence pricing.
- */
-void setConsumablePrices(int size, int *consumablePrices,
-                         Consumable **consumables)
-{
-
-    for (int i = 0; i < size; i++) {
-        int *cur_price = &consumablePrices[i];
-        //Consumable* c = consumables[i];
-
-        //Price evaluation
-        int baseprice = 4;
-        int price = baseprice + (rand() % 5) - 1;
-
-        *cur_price = price;
-    }
-}
-
-/**
- * Takes one Shop pointer and initialises all the fields, depeding on stats from the Fighter pointer passed.
- * @see Shop
- * @param s The Shop whose fields will be initialised.
- * @param indexWeight The integer for room index weight.
- * @param player The Fighter player to influence item generation.
- * @param t_kls The Koliseo_Temp used for allocations.
- */
-void initShop(Shop *s, int indexWeight, Fighter *player, Koliseo_Temp *t_kls)
-{
-
-    int equipsCount = (rand() % EQUIP_SHOP_MAX) + 1;
-    Koliseo *kls = t_kls->kls;
-
-    if (equipsCount != 0) {
-
-        for (int equip_index = 0; equip_index < equipsCount; equip_index++) {
-            int curr = (rand() % (EQUIPSMAX + 1));
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "Prepping Equip (%i/%i) for Shop", equip_index,
-                    equipsCount);
-            kls_log(kls, "DEBUG", "Prepping Equip (%i/%i) for Shop",
-                    equip_index, equipsCount);
-            Equip *e =
-                (Equip *) KLS_PUSH_T_TYPED(t_kls, Equip, HR_Equip, "Equip",
-                                           "Equip");
-
-            //Randomise quality
-            quality q = rand() % (QUALITIESMAX + 1);
-
-            //Get the base item and copy the stats to the current equip
-            Equip *base = &equips[curr];
-
-            e->class = base->class;
-            e->type = base->type;
-            e->qual = q;
-
-            setEquipSprite(e);
-            strcpy(e->name, base->name);
-            strcpy(e->desc, base->desc);
-
-            e->qty = 1;
-            e->equipped = 0;
-            e->perksCount = 0;
-
-            //Calc drop level
-            e->level = base->level + round(player->level / EQUIPLVLBOOSTRATIO);
-
-            //Chance for better leveled item
-            if ((rand() % 8) - (player->luck / 10) <= 0) {	//Should use a defined constant
-                e->level += 1;	//At least a simple +1
-                if ((rand() % 25) - (player->luck / 10) <= 0) {	//Should use a defined constant
-                    e->level += 1;	//A bonus roll for another +1
-
-                }
-            }
-
-            float boostFactor = 0.7;
-
-            float lvlBoost = boostFactor * (e->level - 1);
-
-            e->atk = round((base->atk * 1.0) + lvlBoost);
-            e->def = round((base->def * 1.0) + lvlBoost);
-            e->vel = round((base->vel * 1.0) + lvlBoost);
-            e->enr = round((base->enr * 1.0) + lvlBoost);
-
-            //Bonus stats on better quality items? Simple for now
-            //
-            if (q == Good) {
-                e->atk += (rand() % 3);	//Should use a defined constant
-                e->def += (rand() % 3);	//Should use a defined constant
-                e->vel += (rand() % 3);	//Should use a defined constant
-                e->enr += (rand() % 2);	//Should use a defined constant
-            } else if (q == Bad) {
-                e->atk -= (rand() % 3);	//Should use a defined constant
-                e->def -= (rand() % 3);	//Should use a defined constant
-                e->vel -= (rand() % 3);	//Should use a defined constant
-                e->enr -= (rand() % 2);	//Should use a defined constant
-                if (e->atk < 0) {
-                    e->atk = 0;
-                };
-                if (e->def < 0) {
-                    e->def = 0;
-                };
-                if (e->vel < 0) {
-                    e->vel = 0;
-                };
-                if (e->enr < 0) {
-                    e->enr = 0;
-                };
-            }
-            //Possible perks for the Equip
-
-            for (int i = 0; i < (EQUIPPERKSMAX); i++) {
-                int chance = 35;
-
-                if (q == Good) {
-                    chance *= 1.5;
-                }
-
-                if ((rand() % 100) < chance) {
-
-                    e->perksCount += 1;
-                    log_tag("debug_log.txt", "[DEBUG]",
-                            "Prepping Perk (%i) for Shop Equip (%i/%i)", i,
-                            equip_index, equipsCount);
-                    kls_log(kls, "DEBUG",
-                            "Prepping Perk (%i) for Shop Equip (%i/%i)", i,
-                            equip_index, equipsCount);
-                    Perk *p =
-                        (Perk *) KLS_PUSH_T_TYPED(t_kls, Perk, HR_Perk,
-                                                  "Perk", "Perk");
-                    p->class = rand() % (PERKSMAX + 1);
-                    //p->name = (char*)malloc(sizeof(nameStringFromPerk(p->class)));
-                    strcpy(p->name, nameStringFromPerk(p->class));
-                    //p->desc = (char*)malloc(sizeof(descStringFromPerk(p->class)));
-                    strcpy(p->desc, descStringFromPerk(p->class));
-                    p->innerValue = 1;
-                    e->perks[(e->perksCount - 1)] = p;
-                }
-            }
-
-            //Set value of how many bonuses we have. Although Equip only has ONE function pointer field for now
-            e->bonus = base->bonus;
-            //Randomise if the item will have an effect function.
-            //Not yet implemented
-            //Initialisation of function happens here
-            //
-            //e->equip_fun =    ;
-
-            //Calc cost value
-
-            float cost = 3;
-
-            cost += 4.5 * (e->qual + 1);
-            cost += 7.5 * (e->perksCount);
-
-            cost += 10.2 * (e->level);
-
-            e->cost = floor(cost);
-
-            s->equips[equip_index] = e;
-            s->equipsCount++;
-
-        }			//End for all equips
-    }
-    s->equipsCount = equipsCount;
-    setEquipPrices(s->equipsCount, s->equipPrices, s->equips);
-
-    int uniqueConsumablesCount = (rand() % SINGLE_CONSUMABLE_SHOP_MAX) + 1;
-    int uniques = 0;
-    //int not_uniques = 0;
-    if (uniqueConsumablesCount != 0) {
-        int already_rolled[CONSUMABLESMAX + 1];
-        for (int i = 0; i < (CONSUMABLESMAX + 1); i++) {
-            already_rolled[i] = 0;
-        }
-        int cons_prepared = 0;
-        while (cons_prepared < uniqueConsumablesCount) {
-
-            for (int i = 0; i < CONSUMABLESMAX + 1; i++) {
-                if (cons_prepared < uniqueConsumablesCount) {
-                    int curr = rand() % (CONSUMABLESMAX + 1);
-                    if (!(already_rolled[curr])) {
-                        log_tag("debug_log.txt", "[DEBUG]",
-                                "Prepping Consumable (%i/%i) for Shop",
-                                cons_prepared, uniqueConsumablesCount);
-                        kls_log(kls, "DEBUG",
-                                "Prepping Consumable (%i/%i) for Shop",
-                                cons_prepared, uniqueConsumablesCount);
-                        Consumable *cur =
-                            (Consumable *) KLS_PUSH_T_TYPED(t_kls, Consumable,
-                                                            HR_Consumable,
-                                                            "Consumable",
-                                                            "Consumable");
-                        cur->class = curr;
-                        already_rolled[curr] = 1;
-                        if (uniqueConsumablesCount - cons_prepared > 0) {
-                            cur->qty = (rand() % 4) + 1;
-                        } else {
-                            cur->qty = 1;
-                            //cur->qty = (rand() % (consumablesCount - cons_index)) +1;
-                        }
-                        setConsumableSprite(cur);
-                        //not_uniques += cur->qty;
-                        cons_prepared++;	//= cur->qty;
-                        s->consumables[uniques] = cur;
-                        uniques++;
-                    }
-                }
-            };
-        }
-    }
-    s->consumablesCount = uniqueConsumablesCount;
-    setConsumablePrices(s->consumablesCount, s->consumablePrices,
-                        s->consumables);
-
-    s->itemCount = s->consumablesCount + s->equipsCount;
-}
-
-/**
- * Takes a Chest and Fighter pointers and initialises all the fields in chest based on chest class and fighter stats.
- * @see Chest
- * @param c The allocated Chest pointer with already set class to initialise.
- * @param f The Fighter pointer with stats.
- * @param t_kls The Koliseo_Temp used for allocations.
- */
-void initChest(Chest *c, Fighter *f, Koliseo_Temp *t_kls)
-{
-
-    setChestSprite(c);
-
-    strcpy(c->desc, descStringFromChest(c->class));
-    int cons_cnt, equip_cnt;
-    Koliseo *kls = t_kls->kls;
-
-    switch (c->class) {
-    case CHEST_BASE: {
-        cons_cnt = rand() % (CHEST_CONSUMABLES_MAX - 1);
-        equip_cnt = rand() % (CHEST_EQUIPS_MAX - 1);
-    }
-    break;
-    case CHEST_BEAST: {
-        cons_cnt = (rand() % (CHEST_CONSUMABLES_MAX)) + 1;
-        equip_cnt = (rand() % (CHEST_EQUIPS_MAX)) + 1;
-
-    }
-    break;
-    default: {
-        log_tag("debug_log.txt", "[ERROR]",
-                "%i is not a valid chest class.", c->class);
-        exit(EXIT_FAILURE);
-    }
-    break;
-
-    }
-
-    c->consumablesCount = cons_cnt;
-    c->equipsCount = equip_cnt;
-
-    if (c->consumablesCount > 0) {
-        for (int i = 0; i < c->consumablesCount; i++) {
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "Prepping Consumable (%i/%i) for Chest", i,
-                    c->consumablesCount);
-            kls_log(kls, "DEBUG", "Prepping Consumable (%i/%i) for Chest", i,
-                    c->consumablesCount);
-            Consumable *cns =
-                (Consumable *) KLS_PUSH_T_TYPED(t_kls, Consumable,
-                                                HR_Consumable, "Consumable",
-                                                "Consumable");
-            int drop = rand() % (CONSUMABLESMAX + 1);
-
-            cns->class = drop;
-
-            Consumable *base = &consumablesBase[drop];
-
-            strcpy(cns->name, base->name);
-            strcpy(cns->desc, base->desc);
-            for (int j = 0; j < 8; j++) {
-                strcpy(cns->sprite[j], base->sprite[j]);
-            }
-            cns->qty = 1;
-
-            c->consumables[i] = cns;
-        }
-    }
-
-    if (c->equipsCount > 0) {
-        for (int i = 0; i < c->equipsCount; i++) {
-
-            //Select a basic item from the list
-            int drop = rand() % (EQUIPSMAX + 1);
-            //Randomise quality
-            quality q = rand() % (QUALITIESMAX + 1);
-
-            //Prepare the item
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "Prepping Equip (%i/%i) for Chest", i, c->equipsCount);
-            kls_log(kls, "DEBUG", "Prepping Equip (%i/%i) for Chest", i,
-                    c->equipsCount);
-            Equip *e =
-                (Equip *) KLS_PUSH_T_TYPED(t_kls, Equip, HR_Equip, "Equip",
-                                           "Equip");
-
-            //Get the base item and copy the stats to the drop
-            Equip *base = &equips[drop];
-
-            e->class = base->class;
-            e->type = base->type;
-            e->qual = q;
-
-            setEquipSprite(e);
-            strcpy(e->name, base->name);
-            strcpy(e->desc, base->desc);
-
-            e->qty = 1;
-            e->equipped = 0;
-
-            e->perksCount = 0;
-
-            //Calc drop level
-            e->level = base->level + round(f->level / EQUIPLVLBOOSTRATIO);
-
-            //Chance for better leveled item
-            if ((rand() % 8) - (f->luck / 10) <= 0) {	//Should use a defined constant
-                e->level += 1;	//At least a simple +1
-                if ((rand() % 25) - (f->luck / 10) <= 0) {	//Should use a defined constant
-                    e->level += 1;	//A bonus roll for another +1
-
-                }
-            }
-
-            float boostFactor = 0.7;
-
-            float lvlBoost = boostFactor * (e->level - 1);
-
-            e->atk = round((base->atk * 1.0) + lvlBoost);
-            e->def = round((base->def * 1.0) + lvlBoost);
-            e->vel = round((base->vel * 1.0) + lvlBoost);
-            e->enr = round((base->enr * 1.0) + lvlBoost);
-
-            //Bonus stats on better quality items? Simple for now
-
-            if (q == Good) {
-                e->atk += (rand() % 3);	//Should use a defined constant
-                e->def += (rand() % 3);	//Should use a defined constant
-                e->vel += (rand() % 3);	//Should use a defined constant
-                e->enr += (rand() % 2);	//Should use a defined constant
-            } else if (q == Bad) {
-                e->atk -= (rand() % 3);	//Should use a defined constant
-                e->def -= (rand() % 3);	//Should use a defined constant
-                e->vel -= (rand() % 3);	//Should use a defined constant
-                e->enr -= (rand() % 2);	//Should use a defined constant
-                if (e->atk < 0) {
-                    e->atk = 0;
-                };
-                if (e->def < 0) {
-                    e->def = 0;
-                };
-                if (e->vel < 0) {
-                    e->vel = 0;
-                };
-                if (e->enr < 0) {
-                    e->enr = 0;
-                };
-            }
-            //Possible perks for the Equip
-
-            for (int j = 0; j < (EQUIPPERKSMAX); j++) {
-                int chance = 20;
-
-                if (q == Good) {
-                    chance *= 1.5;
-                }
-
-                if ((rand() % 100) < chance
-                    || (c->class == CHEST_BEAST && e->perksCount == 0)) {
-
-                    e->perksCount += 1;
-
-                    log_tag("debug_log.txt", "[DEBUG]",
-                            "Prepping Perk (%i/%i) for Equip (%i/%i) for Chest",
-                            j, e->perksCount, i, c->equipsCount);
-                    kls_log(kls, "DEBUG",
-                            "Prepping Perk (%i/%i) for Equip (%i/%i) for Chest",
-                            j, e->perksCount, i, c->equipsCount);
-                    Perk *p =
-                        (Perk *) KLS_PUSH_T_TYPED(t_kls, Perk, HR_Perk,
-                                                  "Perk", "Perk");
-                    p->class = rand() % (PERKSMAX + 1);
-                    //p->name = (char*)malloc(sizeof(nameStringFromPerk(p->class)));
-                    strcpy(p->name, nameStringFromPerk(p->class));
-                    //p->desc = (char*)malloc(sizeof(descStringFromPerk(p->class)));
-                    strcpy(p->desc, descStringFromPerk(p->class));
-                    p->innerValue = 1;
-                    e->perks[(e->perksCount - 1)] = p;
-                }
-            }
-
-            //Set value of how many bonuses we have. Although Equip only has ONE function pointer field for now
-            e->bonus = base->bonus;
-            //Randomise if the item will have an effect function.
-            //Not yet implemented
-            //Initialisation of function happens here
-            //
-            //e->equip_fun =    ;
-
-            //Calc cost value
-
-            float cost = 5;
-
-            cost += 2.5 * (e->qual);
-            cost += 3.5 * (e->perksCount);
-
-            cost += 7.2 * (e->level);
-
-            e->cost = floor(cost);
-
-            c->equips[i] = e;
-        }			//End for all equips
-    }				//End if equipsCount > 0
-}
-
-/**
- * Takes a Chest and Fighter pointers and initialises all the fields in chest making it ready for use in battle, using fighter stats to influence init.
- * Calls initChest() after setting class.
- * @see Chest
- * @see initChest()
- * @param c The allocated Chest pointer to initialise.
- * @param f The Fighter pointer to influence item generation.
- * @param t_kls The Koliseo_Temp used for allocations.
- */
-void prepareChest(Chest *c, Fighter *f, Koliseo_Temp *t_kls)
-{
-
-    //Init chest class
-    int drop = (rand() % 100) + 1;
-
-    if (drop > 70) {
-        c->class = CHEST_BASE;
-    } else {
-        c->class = CHEST_BEAST;
-    }
-
-    //Load Chest stats
-    initChest(c, f, t_kls);
-
-}
-
-/**
- * Takes a Treasure and Fighter pointers and initialises all the treasure fields based on its class and fighter's stats.
- * @see Treasure
- * @param t The allocated Treasure pointer with already set class to initialise.
- * @param f The Fighter pointer to influence item generation.
- * @param t_kls The Koliseo_Temp used for allocations.
- */
-void initTreasure(Treasure *t, Fighter *f, Koliseo_Temp *t_kls)
-{
-
-    Koliseo *kls = t_kls->kls;
-    strcpy(t->desc, descStringFromTreasure(t->class));
-
-    switch (t->class) {
-    case TREASURE_CHEST: {
-
-        log_tag("debug_log.txt", "[DEBUG]",
-                "Allocated %lu for Treasure [Chest]:", sizeof(Chest));
-        kls_log(kls, "DEBUG", "Allocated %lu for Treasure [Chest]:",
-                sizeof(Chest));
-        Chest *c =
-            (Chest *) KLS_PUSH_T_TYPED(t_kls, Chest, HR_Chest, "Chest",
-                                       "Chest");
-        prepareChest(c, f, t_kls);
-        t->chest = c;
-
-    }
-    break;
-    case TREASURE_CONSUMABLE: {
-        log_tag("debug_log.txt", "[DEBUG]",
-                "Allocated %lu for Treasure [Consumable]:",
-                sizeof(Consumable));
-        kls_log(kls, "DEBUG", "Allocated %lu for Treasure [Consumable]:",
-                sizeof(Consumable));
-        Consumable *cns =
-            (Consumable *) KLS_PUSH_T_TYPED(t_kls, Consumable,
-                                            HR_Consumable, "Consumable",
-                                            "Consumable");
-        int drop = rand() % (CONSUMABLESMAX + 1);
-
-        cns->class = drop;
-
-        Consumable *base = (Consumable *) f->consumablesBag[drop];
-
-        strcpy(cns->name, base->name);
-        strcpy(cns->desc, base->desc);
-        for (int j = 0; j < 8; j++) {
-            strcpy(cns->sprite[j], base->sprite[j]);
-        }
-        cns->qty = 1;
-
-        t->consumable = cns;
-    }
-    break;
-    case TREASURE_ARTIFACT: {
-        log_tag("debug_log.txt", "[DEBUG]",
-                "Allocated %lu for Treasure [Artifact]:", sizeof(Artifact));
-        kls_log(kls, "DEBUG", "Allocated %lu for Treasure [Artifact]:",
-                sizeof(Artifact));
-        Artifact *a =
-            (Artifact *) KLS_PUSH_T_TYPED(t_kls, Artifact, HR_Artifact,
-                                          "Artifact", "Artifact");
-        int drop = -1;
-        do {
-            drop = rand() % (ARTIFACTSMAX + 1);
-        } while (f->artifactsBag[drop]->qty > 0);
-
-        a->class = drop;
-
-        Artifact *base = f->artifactsBag[drop];
-
-        strcpy(a->name, base->name);
-        strcpy(a->desc, base->desc);
-        for (int j = 0; j < 8; j++) {
-            strcpy(a->sprite[j], base->sprite[j]);
-        }
-        a->qty = 0;
-
-        t->artifact = a;
-
-    }
-    break;
-    }
 }
 
 /**
